ROM-Cache-MEM2 bug fixup (length/4 would round much causing bad reads)

+++ trunk/main/ROM-Cache.h
@@ -33,7 +33,7 @@
 
 void ROMCache_init(fileBrowser_file*);
 void ROMCache_deinit();
-void ROMCache_read(u32* ram_dest, u32 rom_offset, u32 length);
+void ROMCache_read(u8* ram_dest, u32 rom_offset, u32 length);
 int ROMCache_load(fileBrowser_file* file);
 // WARNING: Not necessarily valid after another ROMCache call
 void* ROMCache_pointer(u32 rom_offset);
